All RegistrationController methods now interact with the database for CRUD operations

+++ app/Http/Controllers/RegistrationController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
+use App\User;
+use App\Meeting;
 
 class RegistrationController extends Controller
 {
@@ -21,26 +23,38 @@ class RegistrationController extends Controller
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
 
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        // Get the meeting via meeting ID
+        // Get the user via user ID
+        $meeting = Meeting::findOrFail($meeting_id);
+        $user = User::findOrFail($user_id);
+
+        // Checking if user is already registered for a meeting
+        $message = 
+        [
+            'msg' => 'User is already registered for meeting',
+            'user' => $user,
+            'meeting' => $meeting,
+            'unregister' => [
+                'href' => 'api/v1/meeting/registration/'.$meeting->id,
+                'method' => 'DELETE'
             ]
         ];
 
-        $user = [
-            'name' => 'Name'
-        ];
+        if ($meeting->users()->where('users.id', $user->id)->first())
+        {
+            return response()->json($message, 404);
+        };
 
-        $response = [
+        // User not already registered - registering them
+        $user->meetings()->attach($meeting);
+
+        $response = 
+        [
             'msg' => 'User registered for meeting',
-            'meeting' => $meeting,
             'user' => $user,
+            'meeting' => $meeting,
             'unregister' => [
-                'href' => 'api/v1/meeting/registration/1',
+                'href' => 'api/v1/meeting/registration/'.$meeting->id,
                 'method' => 'DELETE'
             ]
         ];
@@ -56,24 +70,13 @@ class RegistrationController extends Controller
      */
     public function destroy($id)
     {
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
-
-        $user = [
-            'name' => 'Name'
-        ];
+        $meeting = Meeting::findOrFail($id);
+        $meeting->users()->detach();
 
         $response = [
             'msg' => 'User unregistered for meeting',
             'meeting' => $meeting,
-            'user' => $user,
+            'user' => 'tbd',
             'register' => [
                 'href' => 'api/v1/meeting/registration',
                 'method' => 'POST',
@@ -81,6 +84,6 @@ class RegistrationController extends Controller
             ]
         ];
 
-        return response()->json($response, 200);
+        return response()->json($response, 200);    
     }
 }
